Run the checksum from inside the release dir
sha256sum outputs checksums in the format of 'somehash path/to/file'.
When a user downloads `somezipfile` and `somezipfile.sha256`, we want to be
able to run `sha256sum somezipfile.sha256`.  However, to get the correct
path name in the checksum file, we have to run from the downloader's
POV, which means that `somezipfile` is in the current directory.